Add ramp to PID position mode
In position PID mode, when in the READY state, issue command
{"set":"ramp","to":<rate>}
where the units of the rate are rad/s. Rate is parsed as a float.
The ramp will cease upon issue of wait, stop, or hitting automatic
limits. The automatic limits are not increased for ramp for consistency
with the current UI implementation (and an abundance of caution).